#19 [feat] 로그인 되지 않은 익명 유저도 api를 사용 할 수 있도록 SecurityContextProvider에서 익명사용자에 대한 코드를 작성해준다.
