+++ src/main/java/kusitms/backend/auth/jwt/JWTFilter.java
@@ -45,6 +45,7 @@ public class JWTFilter extends OncePerRequestFilter {
         String path = request.getServletPath();
         return path.equals("/") || path.startsWith("/login") || path.startsWith("/public")
                 || path.equals("/api/v1/test-error") || path.equals("/api/v1/health-check")
+                || path.equals("/onboarding") || path.equals("/base")
                 || path.equals("/api/v1/signup") || path.equals("/api/v1/send-code") || path.equals("/api/v1/verify-code")
                 || path.equals("/api/v1/token/re-issue")
                 || path.equals("/api/v1/zones/recommend") || path.equals("/api/v1/profiles")
