+++ src/main/java/kusitms/backend/result/domain/repository/ResultRepository.java
@@ -1,0 +1,7 @@
+package kusitms.backend.result.domain.repository;
+
+import kusitms.backend.result.domain.entity.Result;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ResultRepository extends JpaRepository<Result, Long> {
+}
