+++ src/main/java/kusitms/backend/result/domain/repository/ZoneRepository.java
@@ -1,0 +1,11 @@
+package kusitms.backend.result.domain.repository;
+
+import kusitms.backend.result.domain.entity.Result;
+import kusitms.backend.result.domain.entity.Zone;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ZoneRepository extends JpaRepository<Zone, Long> {
+    List<Zone> findAllByResult(Result result);
+}
