+++ src/main/java/kusitms/backend/result/dto/response/SaveTopRankedZoneResponseDto.java
@@ -1,0 +1,10 @@
+package kusitms.backend.result.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record SaveTopRankedZoneResponseDto(Long resultId) {
+    public static SaveTopRankedZoneResponseDto from(Long resultId) {
+        return new SaveTopRankedZoneResponseDto(resultId);
+    }
+}
