--- src/main/java/kusitms/backend/stadium/domain/entity/Profile.java
@@ -1,38 +1,0 @@
-package kusitms.backend.stadium.domain.entity;
-
-import jakarta.persistence.*;
-import kusitms.backend.stadium.domain.converter.StringListConverter;
-import lombok.*;
-
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-@Table(name = "recommend_profile")
-public class Profile {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "profile_id", nullable = false)
-    private Long id;
-
-    @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    @JoinColumn(name = "result_id", nullable = false)
-    private Result result;
-
-    @Column(nullable = false)
-    private String nickname;
-
-    @Column(nullable = false)
-    private String type;
-
-    @Column(nullable = false)
-    private String explanation;
-
-    @Lob
-    @Convert(converter = StringListConverter.class)
-    private List<String> hashTags;
-}
