--- src/main/java/kusitms/backend/stadium/domain/repository/ResultRepository.java
@@ -1,7 +1,0 @@
-package kusitms.backend.stadium.domain.repository;
-
-import kusitms.backend.stadium.domain.entity.Result;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ResultRepository extends JpaRepository<Result, Long> {
-}
