--- src/main/java/kusitms/backend/stadium/dto/response/SaveTopRankedZoneResponseDto.java
@@ -1,10 +1,0 @@
-package kusitms.backend.stadium.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record SaveTopRankedZoneResponseDto(Long resultId) {
-    public static SaveTopRankedZoneResponseDto from(Long resultId) {
-        return new SaveTopRankedZoneResponseDto(resultId);
-    }
-}
